+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(seplyr)
-
-test_check("seplyr")